requirement.txt
tensorflow-gpu==1.12.0
imageio=2.5.0 
numpy=1.16.3 
pillow=6.0.0
joblib=0.13.2 
scikit-learn=0.21.0 
scipy=1.2.1 
sklearn=0.0
